# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Distributed Job Scheduler API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects for an org
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: orgId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
  /queues:
    get:
      operationId: listQueues
      tags: [queues]
      summary: List queues for a project
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Queue"
    post:
      operationId: createQueue
      tags: [queues]
      summary: Create queue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Queue"
  /queues/{queueId}:
    get:
      operationId: getQueue
      tags: [queues]
      summary: Get queue by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: queueId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Queue"
        "404":
          description: Not found
    patch:
      operationId: updateQueue
      tags: [queues]
      summary: Update queue config
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: queueId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueueUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Queue"
  /queues/{queueId}/pause:
    post:
      operationId: pauseQueue
      tags: [queues]
      summary: Pause a queue
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: queueId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Queue"
  /queues/{queueId}/resume:
    post:
      operationId: resumeQueue
      tags: [queues]
      summary: Resume a paused queue
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: queueId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Queue"
  /queues/{queueId}/stats:
    get:
      operationId: getQueueStats
      tags: [queues]
      summary: Get queue statistics
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: queueId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueStats"
...
  /scheduled-jobs:
    get:
      operationId: listScheduledJobs
      tags: [scheduled-jobs]
      summary: List scheduled job templates
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: queueId
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduledJob"
    post:
      operationId: createScheduledJob
      tags: [scheduled-jobs]
      summary: Create a cron-based scheduled job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledJobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledJob"
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List all registered workers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Worker"
  /workers/{workerId}/heartbeats:
    get:
      operationId: getWorkerHeartbeats
      tags: [workers]
      summary: Get heartbeat history for a worker
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: workerId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkerHeartbeat"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      required: [id, orgId, name, createdAt]
      properties:
        id:
          type: string
        orgId:
          type: string
        name:
          type: string
        createdAt:
          type: ["string", "null"]
    ProjectInput:
      type: object
      required: [orgId, name]
      properties:
        orgId:
          type: string
        name:
          type: string
          minLength: 1
...
    Queue:
      type: object
      required: [id, projectId, name, priority, concurrencyLimit, isPaused, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        priority:
          type: integer
        concurrencyLimit:
          type: integer
        isPaused:
          type: boolean
        defaultRetryPolicyId:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    QueueInput:
      type: object
      required: [projectId, name]
      properties:
        projectId:
          type: string
        name:
          type: string
          minLength: 1
        priority:
          type: integer
        concurrencyLimit:
          type: integer
        defaultRetryPolicyId:
          type: string
    QueueUpdate:
      type: object
      properties:
        priority:
          type: integer
        concurrencyLimit:
          type: integer
        defaultRetryPolicyId:
          type: ["string", "null"]
    QueueStats:
      type: object
      required: [queueId, queued, running, completed, failed, deadLetter, total]
      properties:
        queueId:
          type: string
        queued:
          type: integer
        running:
          type: integer
        completed:
          type: integer
        failed:
          type: integer
        deadLetter:
          type: integer
        total:
          type: integer
...
    Worker:
      type: object
      required: [id, name, status, registeredAt]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, draining, dead]
        lastHeartbeatAt:
          type: ["string", "null"]
        registeredAt:
          type: ["string", "null"]
    WorkerHeartbeat:
      type: object
      required: [id, workerId, timestamp]
      properties:
        id:
          type: string
        workerId:
          type: ["string", "null"]
        timestamp:
          type: ["string", "null"]
        cpuUsage:
          type: ["number", "null"]
        memoryUsage:
          type: ["number", "null"]
        activeJobCount:
          type: ["integer", "null"]
    ScheduledJob:
      type: object
      required: [id, queueId, jobTemplate, cronExpression, isActive, createdAt]
      properties:
        id:
          type: string
        queueId:
          type: string
        jobTemplate:
          type: object
        cronExpression:
          type: string
        nextRunAt:
          type: ["string", "null"]
        lastRunAt:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: ["string", "null"]
    ScheduledJobInput:
      type: object
      required: [queueId, jobTemplate, cronExpression]
      properties:
        queueId:
          type: string
        jobTemplate:
          type: object
        cronExpression:
          type: string
...